requirements.txt
 snowflake-connector-python
 snowflake-sqlalchemy
 apache-airflow-providers-snowflake
 apache-airflow[amazon]
 requests